# Add default_tags attribute to StatsD module