--- a/lib/statsd/instrument.rb
+++ b/lib/statsd/instrument.rb
@@ -284,8 +284,13 @@ module StatsD
     end
   end
 
-  attr_accessor :logger, :default_sample_rate, :prefix, :default_tags
+  attr_accessor :logger, :default_sample_rate, :prefix
   attr_writer :backend
+  attr_reader :default_tags
+
+  def default_tags=(tags)
+    @default_tags = StatsD::Instrument::Metric.normalize_tags(tags)
+  end
 
   def backend
     @backend ||= StatsD::Instrument::Environment.default_backend