--- a/lib/statsd/instrument/metric.rb
+++ b/lib/statsd/instrument/metric.rb
@@ -62,7 +62,7 @@ class StatsD::Instrument::Metric
     @sample_rate = options[:sample_rate] || StatsD.default_sample_rate
     @tags        = StatsD::Instrument::Metric.normalize_tags(options[:tags])
     if StatsD.default_tags
-      @tags      = Array(@tags) | StatsD::Instrument::Metric.normalize_tags(StatsD.default_tags)
+      @tags      = Array(@tags) | StatsD.default_tags
     end
     @metadata    = options.reject { |k, _| [:type, :name, :value, :sample_rate, :tags].include?(k) }
   end