--- a/test/benchmark/default_tags.rb
+++ b/test/benchmark/default_tags.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'statsd-instrument'
 require 'benchmark/ips'
 
@@ -6,19 +8,15 @@ StatsD.logger = Logger.new('/dev/null')
 class Suite
   def warming(*args)
     StatsD.default_tags = if args[0] == "with default tags"
-                            {:first_tag => 'first_value', :second_tag => 'second_value'}
-                          else
-                            nil
-                          end
+      { first_tag: 'first_value', second_tag: 'second_value' }
+    end
     puts "warming with default tags: #{StatsD.default_tags}"
   end
 
   def running(*args)
     StatsD.default_tags = if args[0] == "with default tags"
-                            {:first_tag => 'first_value', :second_tag => 'second_value'}
-                          else
-                            nil
-                          end
+      { first_tag: 'first_value', second_tag: 'second_value' }
+    end
     puts "running with default tags: #{StatsD.default_tags}"
   end
 
@@ -32,13 +30,17 @@ end
 suite = Suite.new
 
 Benchmark.ips do |bench|
-  bench.config(:suite => suite)
+  bench.config(suite: suite)
   bench.report("without default tags") do
-    StatsD.increment('GoogleBase.insert', tags: { :first_tag => 'first_value', :second_tag => 'second_value', :third_tag => 'third_value' })
+    StatsD.increment('GoogleBase.insert', tags: {
+      first_tag: 'first_value',
+      second_tag: 'second_value',
+      third_tag: 'third_value',
+    })
   end
 
   bench.report("with default tags") do
-    StatsD.increment('GoogleBase.insert', tags: { :third_tag => 'third_value' })
+    StatsD.increment('GoogleBase.insert', tags: { third_tag: 'third_value' })
   end
 
   bench.compare!