Implements: #155 

Adds new attribute to `StatsD` module allowing to specify default tags that will be applied to all metrics.