@fabiormoura 2019-09-16
> Can you write a simple benchmark, to compare metric emission performance:
> 
> 1. Before this change
> 2. After this change, with no default tags
> 3. After this change, with default tags set.

I'll write the benchmark once the implementation satisfies. I'm making the change you suggested on the other comment.