@wvanbergen 2019-09-16
> I'll write the benchmark once the implementation satisfies.

Performance is a feature. The implementation only satisfies if the performance is good enough.  :)