@wvanbergen 2019-09-16
Final question, for `StatsD.prefix`, we have a `no_prefix` option to prevent the prefix from being added to the metric name. Do we need the same for the `default_tags`?