@fabiormoura 2019-09-16
> Final question, for `StatsD.prefix`, we have a `no_prefix` option to prevent the prefix from being added to the metric name. Do we need the same for the `default_tags`?

It is up to you. To me not having this for now is fine as well. If someone else needs that as a requirement they could contribute through PR. But if you believe that should be part of this PR, let me know.