@fabiormoura 2019-09-16
> Can you write a simple benchmark, to compare metric emission performance:
> 
> 1. Before this change
> 2. After this change, with no default tags
> 3. After this change, with default tags set.

How to run it:

```
bundle exec ruby test/benchmark/default_tags.rb
```

The results are:

> 1. Before this change

```
Warming up --------------------------------------
without default tags     7.050k i/100ms
Calculating -------------------------------------
without default tags     73.708k (± 2.3%) i/s -    373.650k in   5.072079s
```

> 2. After this change, with no default tags
> 3. After this change, with default tags set.

```
Warming up --------------------------------------
without default tags     7.142k i/100ms
   with default tags     4.560k i/100ms
Calculating -------------------------------------
without default tags     49.446k (± 5.5%) i/s -    249.970k in   5.073286s
   with default tags     48.173k (± 1.5%) i/s -    241.680k in   5.018064s

Comparison:
without default tags:    49445.8 i/s
   with default tags:    48172.8 i/s - same-ish: difference falls within error
```