@fabiormoura 2019-09-16
> > Can you write a simple benchmark, to compare metric emission performance:
> > 
> > 1. Before this change
> > 2. After this change, with no default tags
> > 3. After this change, with default tags set.
> 
> How to run it:
> 
> ```
> bundle exec ruby test/benchmark/default_tags.rb
> ```
> 
> The results are:
> 
> > 1. Before this change
> 
> ```
> Warming up --------------------------------------
> without default tags     5.040k i/100ms
> Calculating -------------------------------------
> without default tags     52.150k (± 1.2%) i/s -    262.080k in   5.026255s
> ```
> 
> > 1. After this change, with no default tags
> > 2. After this change, with default tags set.
> 
> ```
> Warming up --------------------------------------
> warming with default tags:
> without default tags     4.963k i/100ms
> warming with default tags: {:first_tag=>"first_value", :second_tag=>"second_value"}
>    with default tags     4.838k i/100ms
> Calculating -------------------------------------
> running with default tags:
> without default tags     52.946k (± 1.4%) i/s -    268.002k in   5.062786s
> running with default tags: {:first_tag=>"first_value", :second_tag=>"second_value"}
>    with default tags     50.408k (± 1.4%) i/s -    256.414k in   5.087807s
> 
> Comparison:
> without default tags:    52945.9 i/s
>    with default tags:    50408.1 i/s - 1.05x  slower
> ```

What I posted originally was incorrect as the benchmark code wasn't setup properly. I edited my original answer to reflect with correct results.