@wvanbergen 2019-09-16
> It is up to you. To me not having this for now is fine as well. If someone else needs that as a requirement they could contribute through another PR.

That works for me :+1: