@wvanbergen 2019-09-16
We are using set union (the `|` operator) to combine the tags. My guess this involves checking the resulting array for duplicates, which is quite computationally heavy.

Given that we don't actually filter out duplicates for key/value-based tags anyway, I think we can completely forget about removing duplicates and combine the two arrays using array concatenation (`+`) instead. Can you run the benchmark again with that change?