@fabiormoura 2019-09-17
> We are using set union (the `|` operator) to combine the tags. My guess this involves checking the resulting array for duplicates, which is quite computationally heavy.
> 
> Given that we don't actually filter out duplicates for key/value-based tags anyway, I think we can completely forget about removing duplicates and combine the two arrays using array concatenation (`+`) instead. Can you run the benchmark again with that change?

Sorry for not giving you the entire context. The code I benchmarked did not have the change you proposed here: https://github.com/Shopify/statsd-instrument/pull/159#discussion_r324825072.

Let me first make this change and then benchmark it again before deciding on what operator to use `|` vs `+`.