@wvanbergen 2019-09-17
Now you have the benchmark script, it'll be straightforward to test the perf impact of some of these small changes.