@fabiormoura 2019-09-17
Below the benchmark after https://github.com/Shopify/statsd-instrument/pull/159#discussion_r324825072 has been implemented: 

```
Warming up --------------------------------------
warming with default tags:
without default tags     4.255k i/100ms
warming with default tags: ["first_tag:first_value", "second_tag:second_value"]
   with default tags     4.877k i/100ms
Calculating -------------------------------------
running with default tags:
without default tags     49.541k (± 5.0%) i/s -    251.045k in   5.081559s
running with default tags: ["first_tag:first_value", "second_tag:second_value"]
   with default tags     53.646k (± 2.8%) i/s -    268.235k in   5.004186s

Comparison:
   with default tags:    53646.1 i/s
without default tags:    49541.4 i/s - 1.08x  slower
```