@fabiormoura 2019-09-17
The other benchmark taken after changing the operator to `+` to test if it caused any improvements:

```ruby
Warming up --------------------------------------
warming with default tags:
without default tags     4.887k i/100ms
warming with default tags: ["first_tag:first_value", "second_tag:second_value"]
   with default tags     5.341k i/100ms
Calculating -------------------------------------
running with default tags:
without default tags     52.123k (± 2.8%) i/s -    263.898k in   5.067190s
running with default tags: ["first_tag:first_value", "second_tag:second_value"]
   with default tags     56.378k (± 3.8%) i/s -    283.073k in   5.029278s

Comparison:
   with default tags:    56377.9 i/s
without default tags:    52122.6 i/s - 1.08x  slower
```