@wvanbergen 2019-09-17
It's very weird to me that in the last two benchmarks, it's faster to use `default_tags` compared to not using them. That doesn't add up for me :/