@fabiormoura 2019-09-17
> It's very weird to me that in the last two benchmarks, it's faster to use `default_tags` compared to not using them. That doesn't add up for me :/

That is justifiable. Let's walk through that step by step:

### No default tags

The code below emits a metrics containing a hash with 3 tags. For every iteration of the test the hash has is normalized.

```
StatsD.increment('GoogleBase.insert', tags: { :first_tag => 'first_value', :second_tag => 'second_value', :third_tag => 'third_value' })
```

### With default tags
```
StatsD.default_tags = {:first_tag => 'first_value', :second_tag => 'second_value'} # set up only once before the benchmark starts
```
and then 

```
StatsD.increment('GoogleBase.insert', tags: { :third_tag => 'third_value' })
```

The difference now is that the `Hash` passed to the metric has a size of 1 and it gets normalized faster compared to the `Hash` of the scenario without default tags where its size is 3.