@fabiormoura 2019-09-20
@wvanbergen please let me know if you think there is more work to be done here.